requirements.txt
asgiref==3.5.2
autopep8==2.0.0
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
colorlog==6.7.0
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
idna==3.4
pycodestyle==2.9.1
PyJWT==2.6.0
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.12
